requirements.txt
dipy==0.14.0
matplotlib==2.2.2
nibabel==2.3.0
numpy==1.15.0
pandas==0.23.4
pydicom==1.2.2
scikit-image==0.14.0
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
SimpleITK==1.1.0
tabulate==0.8.2
tensorboard==1.9.0
tensorflow-gpu==1.9.0
xlrd==1.1.0
